Fix source file type detection, argument not iterable error
When detecting a .srt file type, we can only rely on the
UnicodeDecodeError exception; but this won't work if we
allowed a non-Unicode file (with "encoding='latin-1'" e.g.)
Work around this by trying video type first.

On Win32, directly using [audio] will produce a "TypeError:
argument of type 'WindowsPath' is not iterable" error. Fix
this by using [str(audio)].

Signed-off-by: Tarnyko <tarnyko@tarnyko.net>